requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
beautifulsoup4==4.11.2
blinker==1.5
certifi==2023.7.22
charset-normalizer==2.1.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
outcome==1.3.0
priority==2.0.0
pydantic==2.4.2
pydantic_core==2.10.1
PySocks==1.7.1
quart==0.18.4
requests==2.28.1
selenium==4.6.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.27.0
tinydb==4.8.0
tomli==2.0.1
trio==0.22.2
trio-websocket==0.11.1
urllib3==1.26.18
uvicorn==0.23.2
vk-api==11.9.9
Werkzeug==3.0.1
wsproto==1.2.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==21.2.0
Automat==20.2.0
Babel==2.9.1
bcrypt==3.2.0
blinker==1.6.2
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.3
cloud-init==23.1.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
exceptiongroup==1.1.1
fail2ban==0.11.2
fastapi==0.98.0
Flask==2.3.2
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httplib2==0.20.2
httpx==0.25.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
magic-filter==1.0.9
MarkupSafe==2.1.3
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.0
openai==0.27.8
pexpect==4.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
python-dotenv==1.0.0
pytz==2022.1
PyYAML==5.4.1
regex==2023.6.3
requests==2.31.0
simplejson==3.17.6
SQLAlchemy==2.0.18
tiktoken==0.4.0
typing_extensions==4.7.0
urllib3==1.26.5
uvicorn==0.22.0
